Make extruder_nr size_t in setExtruder_addPrime
It was size_t or unsigned int in all functions that call this function, so the if statement that checks if it's -1 will never occur. Then we can safely make this size_t here as well.

The extruder number should never be -1 after the settings system has had their go on the input obtained from the front-end.

Contributes to issue CURA-4410.

+++ src/FffGcodeWriter.h
@@ -666,9 +666,9 @@ public:
      * 
      * \param[in] storage where the slice data is stored.
      * \param gcode_layer The initial planning of the gcode of the layer.
-     * \param extruder_nr The extruder to which to switch
+     * \param extruder_nr The extruder to switch to.
      */
-    void setExtruder_addPrime(const SliceDataStorage& storage, LayerPlan& gcode_layer, const int extruder_nr) const;
+    void setExtruder_addPrime(const SliceDataStorage& storage, LayerPlan& gcode_layer, const size_t extruder_nr) const;
 
 private:
     /*!
